+++ jsonargparse/_common.py
@@ -19,7 +19,14 @@ from typing import (  # type: ignore[attr-defined]
 )
 
 from ._namespace import Namespace
-from ._optionals import import_reconplogger, reconplogger_support
+from ._optionals import (
+    get_alias_target,
+    get_annotated_base_type,
+    import_reconplogger,
+    is_alias_type,
+    is_annotated,
+    reconplogger_support,
+)
 from ._type_checking import ArgumentParser
 
 __all__ = [
@@ -97,6 +104,19 @@ def is_generic_class(cls) -> bool:
 
 def get_generic_origin(cls):
     return cls.__origin__ if is_generic_class(cls) else cls
+
+
+def get_unaliased_type(cls):
+    new_cls = cls
+    while True:
+        cur_cls = new_cls
+        if is_annotated(new_cls):
+            new_cls = get_annotated_base_type(new_cls)
+        if is_alias_type(new_cls):
+            new_cls = get_alias_target(new_cls)
+        if new_cls == cur_cls:
+            break
+    return cur_cls
 
 
 def is_dataclass_like(cls) -> bool: