+++ jsonargparse_tests/test_dataclass_like.py
@@ -19,6 +19,7 @@ from jsonargparse._optionals import (
     attrs_support,
     docstring_parser_support,
     pydantic_support,
+    pydantic_supports_field_init,
     set_docstring_parse_options,
     typing_extensions_import,
 )
@@ -27,6 +28,9 @@ from jsonargparse_tests.conftest import (
     get_parser_help,
     skip_if_docstring_parser_unavailable,
 )
+
+annotated = typing_extensions_import("Annotated")
+type_alias_type = typing_extensions_import("TypeAliasType")
 
 # dataclass tests
 
@@ -208,14 +212,15 @@ def test_add_argument_dataclass_type_required_attr(parser):
 
 @dataclasses.dataclass
 class DataInitFalse:
-    p1: str = "-"
-    p2: str = dataclasses.field(init=False)
+    p1: str = dataclasses.field(init=False)
 
 
 def test_dataclass_field_init_false(parser):
     added = parser.add_dataclass_arguments(DataInitFalse, "d")
-    assert added == ["d.p1"]
-    assert parser.get_defaults() == Namespace(d=Namespace(p1="-"))
+    assert added == []
+    assert parser.get_defaults() == Namespace()
+    cfg = parser.parse_args(["--d", "{}"])
+    assert cfg.d == Namespace()
 
 
 @dataclasses.dataclass
@@ -468,9 +473,108 @@ def test_add_class_final(parser):
     pytest.raises(ValueError, lambda: parser.add_class_arguments(FinalClass, "a", default=FinalClass()))
 
 
-# pydantic tests
+if type_alias_type:
+    IntOrString = type_alias_type("IntOrString", Union[int, str])
 
-annotated = typing_extensions_import("Annotated")
+    @dataclasses.dataclass
+    class DataClassWithAliasType:
+        p1: IntOrString  # type: ignore[valid-type]
+
+    def test_bare_alias_type(parser):
+        parser.add_argument("--data", type=IntOrString)
+        help_str = get_parser_help(parser)
+        help_str_lines = [line for line in help_str.split("\n") if "type: IntOrString" in line]
+        assert len(help_str_lines) == 1
+        assert "--data DATA" in help_str_lines[0]
+        cfg = parser.parse_args(["--data=MyString"])
+        assert cfg.data == "MyString"
+        cfg = parser.parse_args(["--data=3"])
+        assert cfg.data == 3
+
+    def test_dataclass_with_alias_type(parser):
+        parser.add_argument("--data", type=DataClassWithAliasType)
+        help_str = get_parser_help(parser)
+        help_str_lines = [line for line in help_str.split("\n") if "type: IntOrString" in line]
+        assert len(help_str_lines) == 1
+        assert "--data.p1 P1" in help_str_lines[0]
+        cfg = parser.parse_args(["--data.p1=MyString"])
+        assert cfg.data.p1 == "MyString"
+        cfg = parser.parse_args(["--data.p1=3"])
+        assert cfg.data.p1 == 3
+
+    @pytest.mark.skipif(not annotated, reason="Annotated is required")
+    def test_annotated_alias_type(parser):
+        parser.add_argument("--data", type=annotated[IntOrString, 1])
+        help_str = get_parser_help(parser)
+        help_str_lines = [line for line in help_str.split("\n") if "type: Annotated[IntOrString, 1]" in line]
+        assert len(help_str_lines) == 1
+        assert "--data DATA" in help_str_lines[0]
+        cfg = parser.parse_args(["--data=MyString"])
+        assert cfg.data == "MyString"
+        cfg = parser.parse_args(["--data=3"])
+        assert cfg.data == 3
+
+    if annotated:
+
+        @dataclasses.dataclass
+        class DataClassWithAnnotatedAliasType:
+            p1: annotated[IntOrString, 1]  # type: ignore[valid-type]
+
+    @pytest.mark.skipif(not annotated, reason="Annotated is required")
+    def test_dataclass_with_annotated_alias_type(parser):
+        parser.add_argument("--data", type=DataClassWithAnnotatedAliasType)
+        help_str = get_parser_help(parser)
+        # The printable field datatype is not uniform across versions.
+        help_str_lines = [line for line in help_str.split("\n") if "type:" in line and "IntOrString" in line]
+        assert len(help_str_lines) == 1
+        assert "--data.p1 P1" in help_str_lines[0]
+        cfg = parser.parse_args(["--data.p1=MyString"])
+        assert cfg.data.p1 == "MyString"
+        cfg = parser.parse_args(["--data.p1=3"])
+        assert cfg.data.p1 == 3
+
+
+# pydantic tests
+if annotated and pydantic_support > 1:
+    import pydantic
+
+    @pydantic.dataclasses.dataclass(frozen=True)
+    class InnerDataClass:
+        a2: int = 1
+
+    @pydantic.dataclasses.dataclass(frozen=True)
+    class NestedAnnotatedDataClass:
+        a1: annotated[InnerDataClass, 1]  # type: ignore[valid-type]
+
+    @pydantic.dataclasses.dataclass(frozen=True)
+    class NestedAnnotatedDataClassWithDefault:
+        a1: annotated[InnerDataClass, 1] = pydantic.fields.Field(default=InnerDataClass())  # type: ignore[valid-type]
+
+    @pydantic.dataclasses.dataclass(frozen=True)
+    class NestedAnnotatedDataClassWithDefaultFactory:
+        a1: annotated[InnerDataClass, 1] = pydantic.fields.Field(default_factory=InnerDataClass)  # type: ignore[valid-type]
+
+    def test_pydantic_nested_annotated_dataclass(parser: ArgumentParser):
+        parser.add_class_arguments(NestedAnnotatedDataClass, "n")
+        cfg = parser.parse_args(["--n", "{}"])
+        assert cfg.n == Namespace(a1=Namespace(a2=1))
+
+    def test_pydantic_annotated_nested_annotated_dataclass(parser: ArgumentParser):
+        parser.add_class_arguments(annotated[NestedAnnotatedDataClass, 1], "n")
+        cfg = parser.parse_args(["--n", "{}"])
+        assert cfg.n == Namespace(a1=Namespace(a2=1))
+
+    def test_pydantic_annotated_nested_annotated_dataclass_with_default(parser: ArgumentParser):
+        parser.add_class_arguments(annotated[NestedAnnotatedDataClassWithDefault, 1], "n")
+        cfg = parser.parse_args(["--n", "{}"])
+        assert cfg.n == Namespace(a1=Namespace(a2=1))
+
+    def test_pydantic_annotated_nested_annotated_dataclass_with_default_factory(parser: ArgumentParser):
+        parser.add_class_arguments(annotated[NestedAnnotatedDataClassWithDefaultFactory, 1], "n")
+        cfg = parser.parse_args(["--n", "{}"])
+        assert cfg.n == Namespace(a1=Namespace(a2=1))
+
+
 length = "length"
 if pydantic_support:
     import pydantic
@@ -479,6 +583,26 @@ if pydantic_support:
     class PydanticData:
         p1: float = 0.1
         p2: str = "-"
+
+    @pydantic.dataclasses.dataclass
+    class PydanticDataNested:
+        p3: PydanticData
+
+    if pydantic_supports_field_init:
+        from pydantic.dataclasses import dataclass as pydantic_v2_dataclass
+        from pydantic.fields import Field as PydanticV2Field
+
+        @pydantic_v2_dataclass
+        class PydanticDataFieldInitFalse:
+            p1: str = PydanticV2Field("-", init=False)
+
+    @pydantic.dataclasses.dataclass
+    class PydanticDataStdlibField:
+        p1: str = dataclasses.field(default="-")
+
+    @pydantic.dataclasses.dataclass
+    class PydanticDataStdlibFieldWithFactory:
+        p1: str = dataclasses.field(default_factory=lambda: "-")
 
     class PydanticModel(pydantic.BaseModel):
         p1: str
@@ -587,6 +711,43 @@ class TestPydantic:
             parser.parse_args([f"--model.param={invalid_value}"])
         ctx.match("model.param")
 
+    @pytest.mark.skipif(not pydantic_supports_field_init, reason="Field.init is required")
+    def test_dataclass_field_init_false(self, parser):
+        # Prior to PR #480, this test would produce the following error:
+        #
+        # TypeError: Parser key "data.p1":
+        # Expected a <class 'str'>. Got value: annotation=str
+        # required=False default='-' init=False
+        parser.add_argument("--data", type=PydanticDataFieldInitFalse)
+        help_str = get_parser_help(parser)
+        assert "--data.p1" not in help_str
+        cfg = parser.parse_args(["--data", "{}"])
+        assert cfg.data == Namespace()
+
+        cfg = parser.instantiate_classes(cfg)
+        assert cfg.data.p1 == "-"
+
+    def test_dataclass_stdlib_field(self, parser):
+        parser.add_argument("--data", type=PydanticDataStdlibField)
+        cfg = parser.parse_args(["--data", "{}"])
+        assert cfg.data == Namespace(p1="-")
+
+    def test_dataclass_stdlib_field_init_with_factory(self, parser):
+        parser.add_argument("--data", type=PydanticDataStdlibFieldWithFactory)
+        cfg = parser.parse_args(["--data", "{}"])
+        assert cfg.data == Namespace(p1="-")
+
+    def test_dataclass_nested(self, parser):
+        # Prior to PR #480, this test would produce the following error:
+        #
+        # ValueError: Expected "default" argument to be an instance of
+        # "PydanticData" or its kwargs dict, given
+        # <dataclasses._MISSING_TYPE object at 0x105624c50>
+
+        parser.add_argument("--data", type=PydanticDataNested)
+        cfg = parser.parse_args(["--data", '{"p3": {"p1": 1.0}}'])
+        assert cfg.data == Namespace(p3=Namespace(p1=1.0, p2="-"))
+
 
 # attrs tests
 
@@ -605,6 +766,13 @@ if attrs_support:
     @attrs.define
     class AttrsFieldFactory:
         p1: List[str] = attrs.field(factory=lambda: ["one", "two"])
+
+    @attrs.define
+    class AttrsFieldInitFalse:
+        p1: dict = attrs.field(init=False)
+
+        def __attrs_post_init__(self):
+            self.p1 = {}
 
 
 @pytest.mark.skipif(not attrs_support, reason="attrs package is required")
@@ -628,3 +796,17 @@ class TestAttrs:
         assert cfg1.data.p1 == ["one", "two"]
         assert cfg1.data.p1 == cfg2.data.p1
         assert cfg1.data.p1 is not cfg2.data.p1
+
+    def test_field_init_false(self, parser):
+        # Prior to PR #480, this test would produce the following error:
+        #
+        # TypeError('Validation failed: Key "data.p1" is required but
+        # not included in config object or its value is None.')
+
+        parser.add_argument("--data", type=AttrsFieldInitFalse)
+        cfg = parser.parse_args(["--data", "{}"])
+        help_str = get_parser_help(parser)
+        assert "--data.p1" not in help_str
+        assert cfg.data == Namespace()
+        cfg = parser.instantiate_classes(cfg)
+        assert cfg.data.p1 == {}